maybe `py` package name is causing confusion in installers.
Projects dependent on this one try to import `vt.utils.commons.commons.py` but I guess they look at `...commons.py` and assume that `.py` is a file extension and maybe that's why they're not able to import this package. Changing `.py` to `.core_py` to see if this error is resolved.